backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
Flask-CORS
python-dotenv
tensorflow
pandas